Inline test fixtures
